ASB-32862 moved streamline_dataset.load_source_dataset to build_dataset.py
